requirements.txt
docker-py
redis
pyyaml
uuid
flask
requests
